chore: product api 페칭 타입 추가
