+++ src/components/pages/main/ProductCarousel.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { EventCarouselType } from '@/types/ResponseDataTypes';
-import ProductlItem from '../../ui/ProductItem';
+import ProductlItem from '../../ui/productItem/ProductItem';
 import useCategoryCarousel from '@/hooks/useCarousel';
 
 export default function ProductCarousel({
