+++ src/types/ResponseDataTypes.ts
@@ -156,25 +156,3 @@ export interface OrderProductType {
   productImageUrl: string;
   discountRate: number;
 }
-
-export interface ProductOptionType {
-  productUuid: string;
-  colorOptionId: number;
-  sizeOptionId: number;
-  stock: number;
-  price: number;
-  discountRate: number;
-}
-
-export interface ProductListDataType {
-  productUuid: string;
-  name: string;
-}
-
-export interface ProductThumbnailDataType {
-  id: number;
-  productUuid: string;
-  thumbnailUrl: string;
-  description: string;
-  defaulted: boolean;
-}
